backend/requirements.txt
fastapi
uvicorn
cognee>=0.1.20
pypdf
python-multipart
python-dotenv
networkx
beautifulsoup4
